New cast flavour for chromecast support
