+++ .gitignore
@@ -35,4 +35,5 @@ captures/
 /controls/proguard-classes.pro
 /.idea/codeStyles/Project.xml
 /.idea/caches/
+/.idea/markdown-navigator/
 /controls/proguard-classes-*
