+++ README.md
@@ -1,2 +1,2 @@
-# OneMobileSDK controls [![Build Status](https://travis-ci.org/aol-public/OneMobileSDK-controls-android.svg?branch=master)](https://travis-ci.org/aol-public/OneMobileSDK-controls-android)
+# OneMobileSDK controls [![Build Status](https://travis-ci.com/VerizonAdPlatforms/OneMobileSDK-controls-android.svg?branch=master)](https://travis-ci.com/VerizonAdPlatforms/OneMobileSDK-controls-android)
 Public Android implementation of the OneMobileSDK default custom controls interface... demonstrating how customers can implement their own custom video player controls.
