+++ build.gradle
@@ -5,7 +5,7 @@
 
 buildscript {
     ext {
-        kotlin_version = '1.3.11'
+        kotlin_version = '1.3.30'
         support_version = '28.0.0'
     }
     repositories {
